@@ -19,38 +19,37 @@ package org.awaitility.test.osgi;
   under the License.
  */
 
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.ops4j.pax.exam.Configuration;
 import org.ops4j.pax.exam.Option;
 import org.ops4j.pax.exam.junit.PaxExam;
 
-import java.util.concurrent.Callable;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.awaitility.Awaitility.await;
-import static org.awaitility.proxy.AwaitilityClassProxy.to;
-import static org.hamcrest.Matchers.is;
+import static org.ops4j.pax.exam.Constants.EXAM_FAIL_ON_UNRESOLVED_KEY;
 import static org.ops4j.pax.exam.CoreOptions.*;
 
-@Ignore
 @RunWith(PaxExam.class)
 public class AwaitilityTest {
 
     @Configuration
-    public static Option[] configure() throws Exception {
-        return new Option[] //
+    public static Option[] configure() {
+        return new Option[]
                 {
-                        mavenBundle("org.apache.servicemix.bundles", "org.apache.servicemix.bundles.hamcrest", "1.3_1"),
-                        junitBundles(),
-                        systemProperty("pax.exam.osgi.unresolved.fail").value("true"),
+                        /* System Properties */
+                        systemProperty(EXAM_FAIL_ON_UNRESOLVED_KEY).value("true"),
                         systemProperty("org.ops4j.pax.logging.DefaultServiceLog.level").value("INFO"),
+
+                        /* Hamcrest & JUnit bundles */
+                        junitBundles(),
+
+                        /* Deps */
+                        mavenBundle().groupId("org.hamcrest").artifactId("hamcrest").versionAsInProject(),
                         mavenBundle("org.awaitility", "awaitility").versionAsInProject(),
-                        mavenBundle("org.objenesis", "objenesis").versionAsInProject(),
-                        mavenBundle("net.bytebuddy", "byte-buddy").versionAsInProject(),
                         // CoreOptions.vmOption("-Xrunjdwp:transport=dt_socket,server=y,suspend=y,address=5005")
                 };
     }
@@ -60,35 +59,10 @@ public class AwaitilityTest {
     @Test
     public void testWaitUntil() {
 
-        Executors.newScheduledThreadPool(1).schedule(new Runnable() {
-            @Override
-            public void run() {
-                success = true;
-            }
+        Executors.newScheduledThreadPool(1).schedule(() -> {
+            success = true;
         }, 1, SECONDS);
-        await().atMost(2, TimeUnit.SECONDS).until(new Callable<Boolean>() {
-            @Override
-            public Boolean call() throws Exception {
-                return success;
-            }
-        });
+        await().atMost(2, TimeUnit.SECONDS).until(() -> success);
 
     }
-
-    @Test
-    public void testProxy() {
-
-        Executors.newScheduledThreadPool(1).schedule(new Runnable() {
-            @Override
-            public void run() {
-                success = true;
-            }
-        }, 1, SECONDS);
-        await().atMost(2, TimeUnit.SECONDS).untilCall(to(this).getState(), is(true));
-    }
-
-    boolean getState() {
-        return success;
-    }
-
 }
