@@ -17,12 +17,17 @@ package org.awaitility;
 
 import org.awaitility.classes.*;
 import org.awaitility.core.ConditionTimeoutException;
+import org.awaitility.core.ForeverDuration;
+import org.hamcrest.Matcher;
+import org.hamcrest.Matchers;
 import org.junit.Before;
 import org.junit.ComparisonFailure;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
+import org.junit.runners.model.TestTimedOutException;
 
+import java.time.Duration;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.Callable;
@@ -31,13 +36,13 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Stream;
 
 import static java.util.concurrent.TimeUnit.*;
 import static org.awaitility.Awaitility.*;
-import static org.awaitility.Duration.ONE_SECOND;
+import static org.awaitility.Durations.*;
 import static org.hamcrest.Matchers.*;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 public class AwaitilityTest {
 
@@ -53,14 +58,14 @@ public class AwaitilityTest {
     }
 
     @Test(timeout = 2000)
-    public void awaitOperationBlocksAutomatically() throws Exception {
+    public void awaitOperationBlocksAutomatically() {
         new Asynch(fakeRepository).perform();
         await().until(fakeRepositoryValueEqualsOne());
         assertEquals(1, fakeRepository.getValue());
     }
 
     @Test(timeout = 2000)
-    public void awaitOperationSupportsSpecifyingPollIntervalUsingTimeunit() throws Exception {
+    public void awaitOperationSupportsSpecifyingPollIntervalUsingTimeunit() {
         new Asynch(fakeRepository).perform();
         with().pollInterval(20, TimeUnit.MILLISECONDS).await().until(fakeRepositoryValueEqualsOne());
         given().pollInterval(20, TimeUnit.MILLISECONDS).await().until(fakeRepositoryValueEqualsOne());
@@ -68,95 +73,95 @@ public class AwaitilityTest {
     }
 
     @Test(timeout = 2000)
-    public void awaitOperationSupportsSpecifyingPollInterval() throws Exception {
+    public void awaitOperationSupportsSpecifyingPollInterval() {
         new Asynch(fakeRepository).perform();
-        with().pollInterval(Duration.ONE_HUNDRED_MILLISECONDS).then().await().until(fakeRepositoryValueEqualsOne());
+        with().pollInterval(ONE_HUNDRED_MILLISECONDS).then().await().until(fakeRepositoryValueEqualsOne());
         assertEquals(1, fakeRepository.getValue());
     }
 
     @Test(timeout = 2000)
-    public void awaitOperationSupportsSpecifyingZeroAsPollDelay() throws Exception {
+    public void awaitOperationSupportsSpecifyingZeroAsPollDelay() {
         new Asynch(fakeRepository).perform();
-        with().pollDelay(Duration.ZERO).pollInterval(Duration.ONE_HUNDRED_MILLISECONDS).then().await().until(fakeRepositoryValueEqualsOne());
+        with().pollDelay(Duration.ZERO).pollInterval(ONE_HUNDRED_MILLISECONDS).then().await().until(fakeRepositoryValueEqualsOne());
         assertEquals(1, fakeRepository.getValue());
     }
 
     @Test(timeout = 2000)
-    public void awaitOperationSupportsSpecifyingZeroAsPollInterval() throws Exception {
+    public void awaitOperationSupportsSpecifyingZeroAsPollInterval() {
         new Asynch(fakeRepository).perform();
-        with().pollDelay(Duration.TWO_HUNDRED_MILLISECONDS).pollInterval(Duration.ZERO).then().await().until(fakeRepositoryValueEqualsOne());
+        with().pollDelay(TWO_HUNDRED_MILLISECONDS).pollInterval(Duration.ZERO).then().await().until(fakeRepositoryValueEqualsOne());
         assertEquals(1, fakeRepository.getValue());
     }
 
     @Test(timeout = 2000)
-    public void awaitOperationDoesntSupportSpecifyingForeverAsPollDelay() throws Exception {
+    public void awaitOperationDoesntSupportSpecifyingForeverAsPollDelay() {
         exception.expect(IllegalArgumentException.class);
         exception.expectMessage("Cannot delay polling forever");
 
         new Asynch(fakeRepository).perform();
-        with().pollDelay(Duration.FOREVER).pollInterval(Duration.ONE_HUNDRED_MILLISECONDS).then().await().until(fakeRepositoryValueEqualsOne());
+        with().pollDelay(ForeverDuration.FOREVER).pollInterval(ONE_HUNDRED_MILLISECONDS).then().await().until(fakeRepositoryValueEqualsOne());
         assertEquals(1, fakeRepository.getValue());
     }
 
     @Test(timeout = 2000)
-    public void awaitOperationDoesntSupportSpecifyingForeverAsPollInterval() throws Exception {
+    public void awaitOperationDoesntSupportSpecifyingForeverAsPollInterval() {
         exception.expect(IllegalArgumentException.class);
         exception.expectMessage("Cannot use a fixed poll interval of length 'forever'");
 
         new Asynch(fakeRepository).perform();
-        with().pollDelay(Duration.ONE_HUNDRED_MILLISECONDS).pollInterval(Duration.FOREVER).then().await().until(fakeRepositoryValueEqualsOne());
+        with().pollDelay(ONE_HUNDRED_MILLISECONDS).pollInterval(ForeverDuration.FOREVER).then().await().until(fakeRepositoryValueEqualsOne());
         assertEquals(1, fakeRepository.getValue());
     }
 
     @Test(timeout = 2000)
-    public void awaitOperationSupportsSpecifyingPollDelay() throws Exception {
+    public void awaitOperationSupportsSpecifyingPollDelay() {
         new Asynch(fakeRepository).perform();
-        with().pollDelay(Duration.ONE_HUNDRED_MILLISECONDS).await().until(fakeRepositoryValueEqualsOne());
+        with().pollDelay(ONE_HUNDRED_MILLISECONDS).await().until(fakeRepositoryValueEqualsOne());
         assertEquals(1, fakeRepository.getValue());
     }
 
     @Test(timeout = 2000L, expected = ConditionTimeoutException.class)
-    public void awaitOperationSupportsDefaultTimeout() throws Exception {
+    public void awaitOperationSupportsDefaultTimeout() {
         Awaitility.setDefaultTimeout(120, TimeUnit.MILLISECONDS);
         await().until(value(), greaterThan(0));
         assertEquals(1, fakeRepository.getValue());
     }
 
     @Test(timeout = 2000)
-    public void foreverConditionSpecificationUsingUntilWithDirectBlock() throws Exception {
+    public void foreverConditionSpecificationUsingUntilWithDirectBlock() {
         new Asynch(fakeRepository).perform();
         await().forever().until(fakeRepositoryValueEqualsOne());
         assertEquals(1, fakeRepository.getValue());
     }
 
     @Test(timeout = 2000)
-    public void foreverConditionWithHamcrestMatchersWithDirectBlock() throws Exception {
+    public void foreverConditionWithHamcrestMatchersWithDirectBlock() {
         new Asynch(fakeRepository).perform();
         await().forever().until(value(), equalTo(1));
         assertEquals(1, fakeRepository.getValue());
     }
 
     @Test(timeout = 2000)
-    public void foreverConditionWithHamcrestCollectionMatchersWithDirectBlock() throws Exception {
+    public void foreverConditionWithHamcrestCollectionMatchersWithDirectBlock() {
         new Asynch(fakeRepository).perform();
         await().forever().until(valueAsList(), hasItem(1));
         assertEquals(1, fakeRepository.getValue());
     }
 
     @Test(timeout = 3000, expected = ConditionTimeoutException.class)
-    public void throwsTimeoutExceptionWhenDoneEarlierThanAtLeastConstraint() throws Exception {
+    public void throwsTimeoutExceptionWhenDoneEarlierThanAtLeastConstraint() {
         new Asynch(fakeRepository).perform();
         await().atLeast(1, SECONDS).and().atMost(2, SECONDS).until(value(), equalTo(1));
     }
 
     @Test(timeout = 3000)
-    public void doesNotThrowTimeoutExceptionWhenDoneLaterThanAtLeastConstraint() throws Exception {
+    public void doesNotThrowTimeoutExceptionWhenDoneLaterThanAtLeastConstraint() {
         new Asynch(fakeRepository).perform();
         await().atLeast(100, NANOSECONDS).until(value(), equalTo(1));
     }
 
     @Test(timeout = 2000)
-    public void specifyingDefaultPollIntervalImpactsAllSubsequentUndefinedPollIntervalStatements() throws Exception {
+    public void specifyingDefaultPollIntervalImpactsAllSubsequentUndefinedPollIntervalStatements() {
         Awaitility.setDefaultPollInterval(20, TimeUnit.MILLISECONDS);
         new Asynch(fakeRepository).perform();
         await().until(value(), equalTo(1));
@@ -164,15 +169,14 @@ public class AwaitilityTest {
     }
 
     @Test(timeout = 2000, expected = ConditionTimeoutException.class)
-    public void conditionBreaksAfterDurationTimeout() throws Exception {
+    public void conditionBreaksAfterDurationTimeout() {
         new Asynch(fakeRepository).perform();
         await().atMost(200, TimeUnit.MILLISECONDS).until(value(), equalTo(1));
         assertEquals(1, fakeRepository.getValue());
     }
 
     @Test(timeout = 2000, expected = IllegalStateException.class)
-    public void uncaughtExceptionsArePropagatedToAwaitingThreadAndBreaksForeverBlockWhenSetToCatchAllUncaughtExceptions()
-            throws Exception {
+    public void uncaughtExceptionsArePropagatedToAwaitingThreadAndBreaksForeverBlockWhenSetToCatchAllUncaughtExceptions() {
         catchUncaughtExceptionsByDefault();
         new ExceptionThrowingAsynch(new IllegalStateException("Illegal state!")).perform();
         await().forever().until(value(), equalTo(1));
@@ -235,7 +239,7 @@ public class AwaitilityTest {
         try {
             given().pollDelay(0, MILLISECONDS).and().ignoreExceptions().await().until(new Callable<Boolean>() {
                 @Override
-                public Boolean call() throws Exception {
+                public Boolean call() {
                     return fakeRepository.getValue() == 3;
                 }
             });
@@ -247,9 +251,24 @@ public class AwaitilityTest {
         }
     }
 
+    @Test(timeout = 2000)
+    public void ignoredExceptionsAreAddedToExceptionHierarchy() {
+        try {
+            await().ignoreExceptions().atMost(200, TimeUnit.MILLISECONDS).until(new Callable<Boolean>() {
+                @Override
+                public Boolean call() throws Exception {
+                    throw new Exception("Nested");
+                }
+            });
+            fail();
+        } catch (ConditionTimeoutException e) {
+            assertNotNull(e.getCause());
+            assertEquals("Nested", e.getCause().getMessage());
+        }
+    }
+
     @Test(timeout = 2000, expected = IllegalStateException.class)
-    public void uncaughtExceptionsArePropagatedToAwaitingThreadAndBreaksForeverBlockWhenCatchingAllUncaughtExceptions()
-            throws Exception {
+    public void uncaughtExceptionsArePropagatedToAwaitingThreadAndBreaksForeverBlockWhenCatchingAllUncaughtExceptions() {
         new ExceptionThrowingAsynch(new IllegalStateException("Illegal state!")).perform();
         catchUncaughtExceptions().and().await().forever().until(value(), equalTo(1));
     }
@@ -283,20 +302,20 @@ public class AwaitilityTest {
             @Override
             public void testLogic() {
                 new ExceptionThrowingAsynch(new IllegalStateException("Illegal state!")).perform();
-                dontCatchUncaughtExceptions().and().await().atMost(Duration.ONE_SECOND).until(value(), equalTo(1));
+                dontCatchUncaughtExceptions().and().await().atMost(ONE_SECOND).until(value(), equalTo(1));
             }
         };
     }
 
     @Test(timeout = 2000, expected = IllegalStateException.class)
-    public void exceptionsInConditionsArePropagatedToAwaitingThreadAndBreaksForeverBlock() throws Exception {
+    public void exceptionsInConditionsArePropagatedToAwaitingThreadAndBreaksForeverBlock() {
         final ExceptionThrowingFakeRepository repository = new ExceptionThrowingFakeRepository();
         new Asynch(repository).perform();
         await().until(new FakeRepositoryValue(repository), equalTo(1));
     }
 
     @Test(timeout = 2000)
-    public void awaitWithAliasDisplaysAliasWhenConditionTimeoutExceptionOccurs() throws Exception {
+    public void awaitWithAliasDisplaysAliasWhenConditionTimeoutExceptionOccurs() {
         String alias = "test";
         exception.expect(ConditionTimeoutException.class);
         exception.expectMessage(
@@ -306,20 +325,21 @@ public class AwaitilityTest {
     }
 
     @Test(timeout = 2000)
-    public void awaitWithAliasDisplaysAliasWhenConditionTimeoutExceptionAndConditionIsACallableOccurs() throws Exception {
+    public void awaitWithAliasDisplaysAliasWhenConditionTimeoutExceptionAndConditionIsACallableOccurs() {
         String alias = "test";
         exception.expect(ConditionTimeoutException.class);
         exception.expectMessage("Condition with alias 'test' didn't complete within 120 milliseconds because condition returned by method \"awaitWithAliasDisplaysAliasWhenConditionTimeoutExceptionAndConditionIsACallableOccurs\" in class org.awaitility.AwaitilityTest was not fulfilled.");
 
+        //noinspection Convert2Lambda - This is because we want to try a real Callable at least once in the test suite
         await(alias).atMost(120, MILLISECONDS).until(new Callable<Boolean>() {
-            public Boolean call() throws Exception {
+            public Boolean call() {
                 return fakeRepository.getValue() > 0;
             }
         });
     }
 
     @Test(timeout = 2000)
-    public void awaitDisplaysSupplierAndMatcherMismatchMessageWhenConditionTimeoutExceptionOccurs() throws Exception {
+    public void awaitDisplaysSupplierAndMatcherMismatchMessageWhenConditionTimeoutExceptionOccurs() {
         exception.expect(ConditionTimeoutException.class);
         exception.expectMessage(FakeRepositoryValue.class.getName()
                 + " expected a value greater than <0> but <0> was equal to <0> within 120 milliseconds.");
@@ -328,7 +348,7 @@ public class AwaitilityTest {
     }
 
     @Test(timeout = 2000)
-    public void awaitDisplaysCallableNameWhenConditionTimeoutExceptionOccurs() throws Exception {
+    public void awaitDisplaysCallableNameWhenConditionTimeoutExceptionOccurs() {
         exception.expect(ConditionTimeoutException.class);
         exception.expectMessage(String.format("Condition %s was not fulfilled within 120 milliseconds.",
                 FakeRepositoryEqualsOne.class.getName()));
@@ -337,8 +357,7 @@ public class AwaitilityTest {
     }
 
     @Test(timeout = 2000)
-    public void awaitDisplaysMethodDeclaringTheCallableWhenCallableIsAnonymousClassAndConditionTimeoutExceptionOccurs()
-            throws Exception {
+    public void awaitDisplaysMethodDeclaringTheCallableWhenCallableIsAnonymousClassAndConditionTimeoutExceptionOccurs() {
         exception.expect(ConditionTimeoutException.class);
         exception
                 .expectMessage(String
@@ -349,8 +368,7 @@ public class AwaitilityTest {
     }
 
     @Test(timeout = 2000)
-    public void awaitDisplaysMethodDeclaringTheSupplierWhenSupplierIsAnonymousClassAndConditionTimeoutExceptionOccurs()
-            throws Exception {
+    public void awaitDisplaysMethodDeclaringTheSupplierWhenSupplierIsAnonymousClassAndConditionTimeoutExceptionOccurs() {
         exception.expect(ConditionTimeoutException.class);
         exception
                 .expectMessage(String
@@ -360,17 +378,43 @@ public class AwaitilityTest {
         with().pollInterval(10, MILLISECONDS).await().atMost(120, MILLISECONDS).until(valueAsAnonymous(), equalTo(2));
     }
 
+    @SuppressWarnings("unchecked")
+    @Test(timeout = 2000)
+    public void awaitDisplaysMethodDeclaringTheSupplierWhenSupplierIsAnonymousClassAndConditionTimeoutExceptionOccursWhenUsingNanos() {
+        exception.expect(ConditionTimeoutException.class);
+        exception.expectMessage(anyOf(Stream.of(equalTo(0).toString(), "null")
+                .map(s -> String.format("%s.valueAsAnonymous Callable expected %s but was %s within 120 nanoseconds.", AwaitilityTest.class.getName(), equalTo(2).toString(), s))
+                .map(Matchers::containsString)
+                .toArray(Matcher[]::new)));
+
+        with().pollInterval(10, NANOSECONDS).await().atMost(120, NANOSECONDS).until(valueAsAnonymous(), equalTo(2));
+    }
+
+    @Test(timeout = 500)
+    public void noArithmeticExceptionIsThrownWhenConvertingASaneAmountOfDaysToNanos() {
+        exception.expect(TestTimedOutException.class); // We don't actually want to wait for 10 days, just that we don't get an ArithmeticException when converting days to nanos
+        with().pollInterval(10, DAYS).await().atMost(120, DAYS).until(valueAsAnonymous(), equalTo(2));
+    }
+
+    @Test(timeout = 2000)
+    public void throwsNiceExceptionWhenPollDurationIsSpecifiedAsNanosAndIsLessThanPollInterval() {
+        exception.expect(IllegalArgumentException.class);
+        exception.expectMessage("Timeout (122 nanoseconds) must be greater than the poll delay (10 milliseconds).");
+
+        with().pollInterval(10, MILLISECONDS).await().atMost(122, NANOSECONDS).until(valueAsAnonymous(), equalTo(2));
+    }
+
     @Test
-    public void awaitilityThrowsIllegalStateExceptionWhenTimeoutIsLessThanPollDelay() throws Exception {
-        exception.expect(IllegalStateException.class);
+    public void awaitilityThrowsIllegalArgumentExceptionWhenTimeoutIsLessThanPollDelay() {
+        exception.expect(IllegalArgumentException.class);
         exception.expectMessage(is("Timeout (10 seconds) must be greater than the poll delay (10 minutes)."));
 
         with().pollDelay(10, MINUTES).await().atMost(10, SECONDS).until(fakeRepositoryValueEqualsOne());
     }
 
     @Test
-    public void awaitilityThrowsIllegalStateExceptionWhenTimeoutIsEqualToPollDelay() throws Exception {
-        exception.expect(IllegalStateException.class);
+    public void awaitilityThrowsIllegalArgumentExceptionWhenTimeoutIsEqualToPollDelay() {
+        exception.expect(IllegalArgumentException.class);
         exception.expectMessage(is("Timeout (200 milliseconds) must be greater than the poll delay (200 milliseconds)."));
 
         with().with().pollDelay(20, MILLISECONDS).pollDelay(200, MILLISECONDS).await().atMost(200, MILLISECONDS).until(fakeRepositoryValueEqualsOne());
@@ -379,12 +423,9 @@ public class AwaitilityTest {
     @Test(timeout = 2000L, expected = IllegalStateException.class)
     public void rethrowsExceptionsInCallable() {
         await().atMost(1, TimeUnit.SECONDS)
-                .until(new Callable<Boolean>() {
-                           public Boolean call() throws Exception {
-                               throw new IllegalStateException("Hello");
-                           }
-                       }
-                );
+                .until(() -> {
+                    throw new IllegalStateException("Hello");
+                });
     }
 
     private Callable<Boolean> fakeRepositoryValueEqualsOne() {
@@ -392,9 +433,10 @@ public class AwaitilityTest {
     }
 
     private Callable<Boolean> fakeRepositoryValueEqualsOneAsAnonymous() {
+        //noinspection Convert2Lambda
         return new Callable<Boolean>() {
 
-            public Boolean call() throws Exception {
+            public Boolean call() {
                 return fakeRepository.getValue() == 1;
             }
         };
@@ -405,18 +447,15 @@ public class AwaitilityTest {
     }
 
     private Callable<Integer> valueAsAnonymous() {
+        //noinspection Convert2Lambda
         return new Callable<Integer>() {
-            public Integer call() throws Exception {
+            public Integer call() {
                 return fakeRepository.getValue();
             }
         };
     }
 
     private Callable<List<Integer>> valueAsList() {
-        return new Callable<List<Integer>>() {
-            public List<Integer> call() throws Exception {
-                return Collections.singletonList(fakeRepository.getValue());
-            }
-        };
+        return () -> Collections.singletonList(fakeRepository.getValue());
     }
 }
