@@ -20,56 +20,53 @@ import org.awaitility.core.ConditionTimeoutException;
 import org.awaitility.core.DeadlockException;
 import org.junit.Test;
 
-import java.util.concurrent.Callable;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.locks.ReentrantLock;
 
 import static org.awaitility.Awaitility.await;
+import static org.awaitility.Durations.ONE_SECOND;
+import static org.awaitility.Durations.TWO_HUNDRED_MILLISECONDS;
 import static org.junit.Assert.*;
 
 public class DeadlockDetectionTest {
 
     @Test(timeout = 2000L)
-    public void deadlockTest() throws Exception {
+    public void deadlockTest() {
         final ReentrantLock lock1 = new ReentrantLock();
         final ReentrantLock lock2 = new ReentrantLock();
 
         final AtomicBoolean locked = new AtomicBoolean();
 
-        final Thread t1 = new Thread() {
-            public void run() {
-                lock1.lock();
+        final Thread t1 = new Thread(() -> {
+            lock1.lock();
+            try {
                 try {
-                    try {
-                        Thread.sleep(50);
+                    Thread.sleep(50);
 
-                        // wait for lock2 or external interrupt() call
-                        lock2.lockInterruptibly();
-                        locked.set(true);
-                    } catch (InterruptedException ignored) {
-                    }
-                } finally {
-                    lock1.unlock();
+                    // wait for lock2 or external interrupt() call
+                    lock2.lockInterruptibly();
+                    locked.set(true);
+                } catch (InterruptedException ignored) {
                 }
+            } finally {
+                lock1.unlock();
             }
-        };
+        });
 
-        final Thread t2 = new Thread(){
-            public void run() {
-                lock2.lock();
+        final Thread t2 = new Thread(() -> {
+            lock2.lock();
+            try {
                 try {
-                    try {
-                        Thread.sleep(50);
+                    Thread.sleep(50);
 
-                        // wait for lock1 or external interrupt() call
-                        lock1.lockInterruptibly();
-                        locked.set(true);
-                    } catch (InterruptedException ignored) {}
-                } finally {
-                    lock2.unlock();
-                }
+                    // wait for lock1 or external interrupt() call
+                    lock1.lockInterruptibly();
+                    locked.set(true);
+                } catch (InterruptedException ignored) {}
+            } finally {
+                lock2.unlock();
             }
-        };
+        });
 
         // start both threads more or less at the same time to create the deadlock
         t1.start();
@@ -78,7 +75,7 @@ public class DeadlockDetectionTest {
         try {
             // wait for something that will never happen
             // i.e. something that will throw the expected ConditionTimeoutException
-            await().atMost(Duration.TWO_HUNDRED_MILLISECONDS).untilTrue(locked);
+            await().atMost(TWO_HUNDRED_MILLISECONDS).untilTrue(locked);
 
             // ... and fail if the exception is not thrown
             fail("ConditionTimeoutException expected.");
@@ -94,11 +91,7 @@ public class DeadlockDetectionTest {
             t2.interrupt();
 
             // wait until both threads are finished
-            await().atMost(Duration.ONE_SECOND).until(new Callable<Boolean>() {
-                public Boolean call() throws Exception {
-                    return !t1.isAlive() && !t2.isAlive();
-                }
-            });
+            await().atMost(ONE_SECOND).until(() -> !t1.isAlive() && !t2.isAlive());
         }
     }
 }
