@@ -19,6 +19,10 @@ package org.awaitility;
 import org.awaitility.pollinterval.IterativePollInterval;
 import org.junit.Test;
 
+import java.time.Duration;
+
+import static org.awaitility.Durations.FIVE_HUNDRED_MILLISECONDS;
+import static org.awaitility.Durations.ONE_SECOND;
 import static org.awaitility.pollinterval.IterativePollInterval.iterative;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
@@ -28,36 +32,36 @@ public class IterativePollIntervalTest {
     @Test public void
     iterative_poll_interval_allows_specifying_start_duration() {
         // Given
-        IterativePollInterval pollInterval = new IterativePollInterval(prev -> prev.multiply(2), Duration.ONE_MILLISECOND);
+        IterativePollInterval pollInterval = new IterativePollInterval(prev -> prev.multipliedBy(2));
 
         // When
         Duration duration = pollInterval.next(1, Duration.ZERO);
 
         // Then
-        assertThat(duration.getValueInMS(), is(2L));
+        assertThat(duration.toMillis(), is(2L));
     }
 
     @Test public void
     iterative_poll_interval_use_no_start_value_by_default() {
         // Given
-        IterativePollInterval pollInterval = new IterativePollInterval(prev -> prev.multiply(2));
+        IterativePollInterval pollInterval = new IterativePollInterval(prev -> prev.multipliedBy(2));
 
         // When
-        Duration duration = pollInterval.next(1, Duration.ONE_SECOND);
+        Duration duration = pollInterval.next(1, Duration.ofSeconds(1));
 
         // Then
-        assertThat(duration.getValueInMS(), is(2000L));
+        assertThat(duration.toMillis(), is(2000L));
     }
 
     @Test public void
     iterative_poll_interval_allows_specifying_start_duration_through_dsl() {
         // Given
-        IterativePollInterval pollInterval = iterative(prev -> prev.multiply(2)).with().startDuration(Duration.FIVE_HUNDRED_MILLISECONDS);
+        IterativePollInterval pollInterval = iterative(prev -> prev.multipliedBy(2)).with().startDuration(FIVE_HUNDRED_MILLISECONDS);
 
         // When
-        Duration duration = pollInterval.next(1, Duration.ONE_SECOND);
+        Duration duration = pollInterval.next(1, ONE_SECOND);
 
         // Then
-        assertThat(duration.getValueInMS(), is(1000L));
+        assertThat(duration.toMillis(), is(1000L));
     }
 }
