@@ -16,8 +16,11 @@
 
 package org.awaitility.pollinterval;
 
-import org.awaitility.Duration;
-import org.awaitility.core.Function;
+
+import java.time.Duration;
+import java.util.function.Function;
+
+import static org.awaitility.core.ForeverDuration.isForever;
 
 /**
  * A poll interval that is generated by a function and a start duration. The function is free to do anything with the duration.
@@ -65,7 +68,7 @@ public class IterativePollInterval implements PollInterval {
         }
         if (startDurationExplicitlyDefined && startDuration == null) {
             throw new IllegalArgumentException("Start duration cannot be null");
-        } else if (startDurationExplicitlyDefined && startDuration.isForever()) {
+        } else if (startDurationExplicitlyDefined && isForever(startDuration)) {
             throw new IllegalArgumentException("Cannot use a poll interval of length 'forever'");
         }
         this.function = function;
